langbridge/requirements-dev.txt
fastapi[standard]==0.117.1
uvicorn[standard]==0.32.0
dependency-injector==4.48.2
sqlalchemy==2.0.43
pydantic-settings==2.11.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
PyYAML==6.0.2
authlib==1.3.1
psycopg[binary]==3.2.3
itsdangerous==2.2.0 
snowflake-connector-python==3.18.0
cryptography==46.0.0
sqlglot==27.25.2
openai==2.2.0
anthropic>=0.39.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain==0.3.27
mcp==1.17.0
